test(aes): separate basic and authenticated tests
