+++ kipher-aes/src/test/kotlin/io/github/jhdcruz/kipher/aes/AesAuthenticatedTest.kt
@@ -1,53 +1,23 @@
+/*
+ * Copyright 2023 Joshua Hero Dela Cruz
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package io.github.jhdcruz.kipher.aes
 
+import io.github.jhdcruz.kipher.aes.AesParams.aad
+import io.github.jhdcruz.kipher.aes.AesParams.decodeToString
+import io.github.jhdcruz.kipher.aes.AesParams.invalidKey
+import io.github.jhdcruz.kipher.aes.AesParams.message
 import io.github.jhdcruz.kipher.common.KipherException
-import org.junit.jupiter.api.RepeatedTest
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertNotEquals
 
-internal class AesEncryptionTest {
-    private val message = "message".encodeToByteArray()
-    private val aad = "metadata".encodeToByteArray()
-    private val invalidKey = "invalid-key".encodeToByteArray()
-
-    private val decodeToString = { bytes: ByteArray -> String(bytes, Charsets.UTF_8) }
-
-    @Test
-    fun `test key generation`() {
-        val cbcEncryption = CbcEncryption()
-
-        // no need to test for authenticated since it uses the same function
-        val key = cbcEncryption.generateKey()
-
-        // 32 = 256-bit key. 16 = 128-bit
-        assertEquals(32, key.size)
-    }
-
-    @RepeatedTest(5)
-    fun `test key randomization`() {
-        val cbcEncryption = CbcEncryption()
-
-        // no need to test for authenticated since it uses the same function
-        val firstKey = cbcEncryption.generateKey()
-        val secondKey = cbcEncryption.generateKey()
-
-        assertNotEquals(firstKey, secondKey)
-    }
-
-    @Test
-    fun `test basic encryption`() {
-        val cbcEncryption = CbcEncryption()
-
-        val encrypted = cbcEncryption.encrypt(message)
-        val decrypted = cbcEncryption.decrypt(encrypted)
-
-        assertEquals(decodeToString(message), decodeToString(decrypted))
-    }
+internal class AesAuthenticatedTest {
 
     @Test
     fun `test authenticated encryption`() {
@@ -60,7 +30,10 @@ internal class AesEncryptionTest {
 
         val decrypted = gcmEncryption.decrypt(encrypted)
 
-        assertEquals(decodeToString(message), decodeToString(decrypted))
+        assertEquals(
+            decodeToString(message),
+            decodeToString(decrypted)
+        )
     }
 
     @Test
@@ -74,7 +47,10 @@ internal class AesEncryptionTest {
             key = encrypted["key"]!!,
         )
 
-        assertEquals(decodeToString(message), decodeToString(decrypted))
+        assertEquals(
+            decodeToString(message),
+            decodeToString(decrypted)
+        )
     }
 
     @Test
@@ -125,30 +101,6 @@ internal class AesEncryptionTest {
                 key = encrypted.getValue("key"),
             )
         }
-    }
-
-    @Test
-    fun `test basic encryption with invalid secret key`() {
-        val cbcEncryption = CbcEncryption()
-
-        assertThrows<KipherException> {
-            cbcEncryption.encrypt(message, invalidKey)
-        }
-    }
-
-    @Test
-    fun `test encryption with custom key size`() {
-        val cbcEncryption = CbcEncryption()
-
-        val secretKey = cbcEncryption.generateKey(128)
-        val encrypted = cbcEncryption.encrypt(message, secretKey)
-
-        val decrypted = cbcEncryption.decrypt(
-            encrypted = encrypted["data"]!!,
-            key = encrypted["key"]!!,
-        )
-
-        assertEquals(decodeToString(message), decodeToString(decrypted))
     }
 
     @ParameterizedTest
