+++ kipher-aes/src/test/kotlin/io/github/jhdcruz/kipher/aes/AesParams.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2023 Joshua Hero Dela Cruz
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.github.jhdcruz.kipher.aes
+
+/**
+ * Sample AES parameters for testing.
+ */
+internal object AesParams {
+    val message = "message".encodeToByteArray()
+    val aad = "metadata".encodeToByteArray()
+    val invalidKey = "invalid-key".encodeToByteArray()
+
+    val decodeToString = { bytes: ByteArray -> String(bytes, Charsets.UTF_8) }
+}
